Adicionando chamada no service da função de exclusão de Employee

+++ src/main/java/com/br/testedev/cadcolab/services/EmployeeService.java
@@ -7,6 +7,7 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import com.br.testedev.cadcolab.Entity.Employee;
+import com.br.testedev.cadcolab.exceptions.EmployeeWithSubordinatesException;
 import com.br.testedev.cadcolab.repositories.IEmployeeRepository;
 
 @Service
@@ -29,7 +30,11 @@ public class EmployeeService {
 	}
 	
 	public void deleteEmployee(Employee employee) {
-		repository.delete(employee);
+		if(!employee.getSubordinates().isEmpty()) {
+			repository.delete(employee);			
+		} else {
+			throw new EmployeeWithSubordinatesException();
+		}
 	}
 
 }
